v5/requirements.txt
# ClaimGuard-CXR v5 pinned dependencies.
torch==2.4.0
torchvision==0.19.0
transformers==4.44.2
open-clip-torch==2.26.1
sentencepiece==0.2.0
accelerate==0.33.0
datasets==2.20.0
peft==0.12.0
trl==0.9.6
faiss-cpu==1.8.0
rank-bm25==0.2.2
wandb==0.17.7
hydra-core==1.3.2
omegaconf==2.3.0
optuna==3.6.1
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
pyarrow==17.0.0
pillow==10.4.0
nibabel==5.2.1
pydicom==2.4.4
presidio-analyzer==2.2.356
presidio-anonymizer==2.2.356
anthropic==0.34.2
openai==1.45.0
scipy==1.14.1
pyyaml==6.0.2
torchxrayvision==1.3.5
scikit-image==0.24.0
